+++ test/project_types/theme/commands/serve_test.rb
@@ -11,7 +11,7 @@ module Theme
         context = ShopifyCLI::Context.new
         ShopifyCLI::Theme::DevServer
           .expects(:start)
-          .with(context, ".", http_bind: Theme::Command::Serve::DEFAULT_HTTP_HOST)
+          .with(context, ".", host: Theme::Command::Serve::DEFAULT_HTTP_HOST)
 
         Theme::Command::Serve.new(context).call
       end
@@ -20,7 +20,7 @@ module Theme
         context = ShopifyCLI::Context.new
         ShopifyCLI::Theme::DevServer
           .expects(:start)
-          .with(context, ".", http_bind: Theme::Command::Serve::DEFAULT_HTTP_HOST)
+          .with(context, ".", host: Theme::Command::Serve::DEFAULT_HTTP_HOST)
           .raises(ShopifyCLI::Theme::DevServer::AddressBindingError)
 
         assert_raises ShopifyCLI::Abort do
@@ -30,17 +30,17 @@ module Theme
 
       def test_can_specify_bind_address
         context = ShopifyCLI::Context.new
-        ShopifyCLI::Theme::DevServer.expects(:start).with(context, ".", http_bind: "0.0.0.0")
+        ShopifyCLI::Theme::DevServer.expects(:start).with(context, ".", host: "0.0.0.0")
 
         command = Theme::Command::Serve.new(context)
-        command.options.flags[:http_bind] = "0.0.0.0"
+        command.options.flags[:host] = "0.0.0.0"
         command.call
       end
 
       def test_can_specify_port
         context = ShopifyCLI::Context.new
         ShopifyCLI::Theme::DevServer.expects(:start).with(context, ".",
-          http_bind: Theme::Command::Serve::DEFAULT_HTTP_HOST, port: 9293)
+          host: Theme::Command::Serve::DEFAULT_HTTP_HOST, port: 9293)
 
         command = Theme::Command::Serve.new(context)
         command.options.flags[:port] = 9293
@@ -50,7 +50,7 @@ module Theme
       def test_can_specify_poll
         context = ShopifyCLI::Context.new
         ShopifyCLI::Theme::DevServer.expects(:start).with(context, ".",
-          http_bind: Theme::Command::Serve::DEFAULT_HTTP_HOST, poll: true)
+          host: Theme::Command::Serve::DEFAULT_HTTP_HOST, poll: true)
 
         command = Theme::Command::Serve.new(context)
         command.options.flags[:poll] = true